# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MUMU API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/sign_up": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "註冊帳號",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "註冊帳號",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "信箱"
                },
                "password": {
                  "type": "string",
                  "example": "密碼"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "電話號碼"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "example@test.com"
                },
                "password": {
                  "type": "string",
                  "example": "example"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/user/sign_in": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/sign_out": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/check": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/protected": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}